requirements.txt
# Core ML and Audio Processing
torch>=2.7.0
torchaudio>=2.7.0
torchvision==0.22.0+cu118
transformers>=4.30.0
librosa>=0.10.0
soundfile>=0.12.0
numpy>=1.21.0

# VAD-specific models
webrtcvad>=2.0.10
faster-whisper>=0.9.0
hear21passt>=0.0.26
torchlibrosa>=0.1.0
torchsummary>=1.5.1

# Data processing
pandas>=1.5.0
scipy>=1.9.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.1.0

# Utilities
PyYAML>=6.0
tqdm>=4.64.0
einops>=0.8.0
huggingface-hub>=0.15.0
datasets>=2.0.0
accelerate>=0.20.0
